backend/pyproject.toml
[project]
name = "evaluation-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "cachetools>=6.2.4",
    "cryptography>=46.0.5",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "instructor[anthropic]>=1.14.4",
    "litellm>=1.81.0",
    "openai>=2.15.0",
    "orjson>=3.11.5",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.45",
    "starlette>=0.50.0",
    "uvicorn>=0.40.0",
    "yalc>=0.1.2",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
    "pytest~=8.4.2",
    "pytest-asyncio>=1.3.0",
    "types-cachetools>=6.2.0.20251022",
]


[tool.mypy]
plugins = [
    "pydantic.mypy"
]

[tool.ruff]
line-length = 70
